+++ noodles-sam/examples/sam_filter_data.rs
@@ -33,7 +33,7 @@ fn main() -> io::Result<()> {
 
     writer.write_header(&header)?;
 
-    for result in reader.records(&header) {
+    for result in reader.record_bufs(&header) {
         let record = result?;
 
         if is_unique_record(&record)? {