+++ noodles-sam/examples/sam_view.rs
@@ -18,7 +18,7 @@ fn main() -> io::Result<()> {
     let stdout = io::stdout().lock();
     let mut writer = sam::Writer::new(BufWriter::new(stdout));
 
-    for result in reader.records(&header) {
+    for result in reader.record_bufs(&header) {
         let record = result?;
         writer.write_record(&header, &record)?;
     }