+++ noodles-sam/src/async/reader.rs
@@ -197,14 +197,14 @@ async fn read_record<R>(
 where
     R: AsyncBufRead + Unpin,
 {
-    use crate::reader::record::parse_record;
+    use crate::reader::record_buf::parse_record_buf;
 
     buf.clear();
 
     match read_line(reader, buf).await? {
         0 => Ok(0),
         n => {
-            parse_record(buf, header, record)
+            parse_record_buf(buf, header, record)
                 .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
             Ok(n)